requirements_api.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Law vector DB
psycopg2-binary==2.9.9
openai==1.30.0
python-dotenv==1.0.1

# Scraping
selenium==4.22.0
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3

# Data
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4

# ML
tensorflow==2.16.2
sentence-transformers==3.0.*
scikit-learn==1.4.2
huggingface_hub==0.22.*

# NLP
spacy>=3.8.3,<3.9
spacy-curated-transformers>=0.3.0
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
